Move admin_client and model_namespace fixtures and remove unnecessary minio fixture
